requirements.txt
# Face Attendance System - Requirements
# Python 3.9.23 Compatible Version

# Core Computer Vision
opencv-python==4.8.1.78

# Image Processing
Pillow>=8.0.0,<10.0.0
numpy>=1.20.0,<1.25.0

# Face Recognition
# Compatible versions for Python 3.9
face-recognition==1.3.0
dlib>=19.24.0

# GUI Framework
# tkinter is part of the Python standard library, not installed via pip

# Network Communication (for mobile camera support)
requests>=2.25.0,<3.0.0

# Additional AI/ML Libraries
scikit-learn>=1.0.0,<1.3.0
scipy>=1.7.0,<1.10.0

# Deep learning for anti-spoofing (Python 3.9 compatible)
# TensorFlow 2.13+ supports Python 3.9 on ARM64 Mac
tensorflow>=2.13.0,<2.16.0
keras>=2.13.0,<2.16.0

# Data Processing
pandas>=1.3.0,<1.6.0
matplotlib>=3.4.0,<3.6.0

# Utility Libraries
python-dateutil>=2.8.0,<3.0.0
setuptools>=40.0.0

# Development and Testing (optional)
pytest>=6.0.0,<8.0.0
flake8>=3.8.0,<7.0.0

# Performance (optional)
numba>=0.53.0,<0.58.0

# Additional Camera Support (optional)
imageio>=2.9.0,<2.32.0
imageio-ffmpeg>=0.4.0,<0.5.0
